+++ packages/ckeditor5-dev-tests/lib/utils/transformfileoptiontotestglob.js
@@ -5,96 +5,82 @@
 
 'use strict';
 
+const path = require( 'path' );
+
 /**
- * Converts values of --files argument to proper globs.
- * There are 5 supported types of values now:
+ * Converts values of `--files` argument to proper globs. These are the supported types of values:
+ *  * "ckeditor5" - matches all root package tests.
+ *  * "*" - matches all packages' files.
+ *  * "foo"  - matches all tests from a package.
+ *  * "!foo" - matches all tests except from a package.
+ *  * "foo/bar/" - matches all tests from a package and a subdirectory.
+ *  * "foo/bar" - matches all tests from a package (or root) with specific filename.
  *
- * 0. the main repository - 'ckeditor5'
- * 1. all packages' files – '*'
- * 2. given package files – 'engine'
- * 3. everything except the given package – '!engine'
- * 4. path – 'engine/view' -> 'ckeditor5-engine/tests/view/**\/*.js'
- * 5. simplified glob – 'engine/view/**\/*.js' -> 'ckeditor5-engine/tests/view/**\/*.js'
- *
- * @param {String} globPattern A path or pattern to determine the tests to execute.
+ * @param {String} pattern A path or pattern to determine the tests to execute.
  * @param {Boolean} [isManualTest=false] Whether the tests are manual or automated.
- * @returns {Iterable.<String>}
+ * @returns {Array.<String>}
  */
-module.exports = function transformFileOptionToTestGlob( globPattern, isManualTest = false ) {
-	const transformedPath = transformSingleGlobPattern( globPattern, { isManualTest } );
-	const transformedPathWithCKEditorPrefix = transformSingleGlobPattern( globPattern, { isManualTest, useCKEditorPrefix: true } );
-	const transformedPathForExternalPackages = transformSingleGlobPattern( globPattern, { isManualTest, externalPackages: true } );
-	const transformedPathForExternalPackagesWithCKEditorPrefix = transformSingleGlobPattern( globPattern, {
+module.exports = function transformFileOptionToTestGlob( pattern, isManualTest = false ) {
+	// Directory to look: `packages/ckeditor5-*/...`.
+	const transformedPath = transformSinglePattern( pattern, { isManualTest } );
+
+	// Directory to look: `packages/ckeditor-*/...`.
+	const transformedPathWithCKEditorPrefix = transformSinglePattern( pattern, { isManualTest, useCKEditorPrefix: true } );
+
+	// Directory to look: `external/*/packages/ckeditor5-*/...`.
+	const transformedPathForExternalPackages = transformSinglePattern( pattern, { isManualTest, externalPackages: true } );
+
+	// Directory to look: `external/*/packages/ckeditor-*/...`.
+	const transformedPathForExternalPackagesWithCKEditorPrefix = transformSinglePattern( pattern, {
 		isManualTest,
 		externalPackages: true,
 		useCKEditorPrefix: true
 	} );
 
-	if (
-		transformedPath === transformedPathWithCKEditorPrefix &&
-		transformedPathForExternalPackages === transformedPathForExternalPackagesWithCKEditorPrefix &&
-		transformedPath === transformedPathForExternalPackages
-	) {
-		return [ transformedPath ];
-	}
-
+	// Return only unique records.
 	return [
-		transformedPath,
-		transformedPathWithCKEditorPrefix,
-		transformedPathForExternalPackages,
-		transformedPathForExternalPackagesWithCKEditorPrefix
+		...new Set( [
+			transformedPath,
+			transformedPathWithCKEditorPrefix,
+			transformedPathForExternalPackages,
+			transformedPathForExternalPackagesWithCKEditorPrefix
+		] )
 	];
 };
 
 /**
- * @param {String} globPattern
+ * @param {String} pattern
  * @param {Object} [options={}]
  * @param {Boolean} [options.isManualTest=false] Whether the tests are manual or automated.
  * @param {Boolean} [options.useCKEditorPrefix=false] If true, the returned path will use 'ckeditor' prefix instead of 'ckeditor5'.
  * @param {Boolean} [options.externalPackages] If true, the returned path will contain "external\/**\/packages".
  * @returns {String}
  */
-function transformSingleGlobPattern( globPattern, options ) {
-	const isManualTest = options.isManualTest || false;
-	const useCKEditorPrefix = options.useCKEditorPrefix || false;
-	const prefix = useCKEditorPrefix ? 'ckeditor' : 'ckeditor5';
+function transformSinglePattern( pattern, options ) {
+	const chunks = pattern.match( /[a-z1-9|*-]+/g );
+	const output = [];
 
-	const globSep = '/';
-	const cwdChunks = process.cwd().split( require( 'path' ).sep );
-	const chunks = globPattern.split( globSep );
-	const packageName = chunks.shift();
-	const globSuffix = [ 'tests', '**' ];
-	let returnChunks = cwdChunks.concat( options.externalPackages ? [ 'external', '*', 'packages' ] : [ 'packages' ] );
+	const isExclusionPattern = pattern.startsWith( '!' );
+	const isFilenamePattern = pattern.includes( '/' ) && !pattern.endsWith( '/' );
 
-	if ( isManualTest ) {
-		globSuffix.push( 'manual', '**' );
+	const prefix = options.useCKEditorPrefix ? 'ckeditor' : 'ckeditor5';
+	const packagesDirectory = options.externalPackages ? [ 'external', '*', 'packages' ] : [ 'packages' ];
+	const packageName = isExclusionPattern ? `!(${ chunks.shift() })*` : chunks.shift();
+	const filename = isFilenamePattern ? chunks.pop() : '*';
+
+	output.push( ...process.cwd().split( path.sep ) );
+
+	if ( packageName !== 'ckeditor5' ) {
+		output.push( ...packagesDirectory, `${ prefix }-${ packageName }` );
 	}
 
-	globSuffix.push( '*.js' );
+	output.push( 'tests' );
 
-	// 0.
-	if ( globPattern === 'ckeditor5' ) {
-		returnChunks = cwdChunks.concat( globSuffix );
-	} else if ( chunks.length === 0 ) {
-		// 1.
-		if ( packageName == '*' ) {
-			returnChunks.push( prefix + '-*', ...globSuffix );
-		} else if ( packageName.startsWith( '!' ) ) {
-			// 3.
-			returnChunks.push( prefix + '-!(' + packageName.slice( 1 ) + ')*', ...globSuffix );
-		} else {
-			// 2.
-			returnChunks.push( prefix + '-' + packageName, ...globSuffix );
-		}
-	} else {
-		// 5.
-		returnChunks.push( prefix + '-' + packageName, 'tests', ...chunks );
-
-		if ( !chunks[ chunks.length - 1 ].endsWith( '.js' ) ) {
-			// 4.
-			returnChunks.push( '**', '*.js' );
-		}
+	if ( options.isManualTest ) {
+		output.push( 'manual' );
 	}
 
-	return returnChunks.join( globSep );
+	output.push( ...chunks, '**', `${ filename }.js` );
+
+	return output.join( path.posix.sep );
 }