+++ packages/ckeditor5-dev-tests/tests/utils/automated-tests/parsearguments.js
@@ -52,18 +52,21 @@ describe( 'parseArguments()', () => {
 			'--theme-path',
 			'/path/to/theme/package',
 			'--additional-languages',
-			'de,fr'
+			'de,fr',
+			'--resolve-js-first'
 		] );
 
 		expect( options[ 'source-map' ] ).to.be.undefined;
 		expect( options[ 'identity-file' ] ).to.be.undefined;
 		expect( options[ 'theme-path' ] ).to.be.undefined;
 		expect( options[ 'additional-languages' ] ).to.be.undefined;
+		expect( options[ 'resolve-js-first' ] ).to.be.undefined;
 
 		expect( options.sourceMap ).to.equal( true );
 		expect( options.identityFile ).to.equal( '/home/.secret/file.key' );
 		expect( options.themePath ).to.equal( '/path/to/theme/package' );
 		expect( options.additionalLanguages ).to.deep.equal( [ 'de', 'fr' ] );
+		expect( options.resolveJsFirst ).to.equal( true );
 	} );
 
 	it( 'deletes all aliases keys from returned object', () => {